+++ node/risk-app/server/routes/__tests__/postcode.js
@@ -1,0 +1,44 @@
+const STATUS_CODES = require('http2').constants
+const createServer = require('../../../server')
+const { mockOptions, mockSearchOptions } = require('../../../test/mock')
+let server, cookie
+
+jest.mock('../../config')
+jest.mock('../../services/flood')
+jest.mock('../../services/address')
+
+beforeAll(async () => {
+  server = await createServer()
+  await server.initialize()
+  const initial = mockOptions()
+
+  const homepageresponse = await server.inject(initial)
+  expect(homepageresponse.statusCode).toEqual(STATUS_CODES.HTTP_STATUS_OK)
+  cookie = homepageresponse.headers['set-cookie'][0].split(';')[0]
+})
+
+afterAll(async () => {
+  await server.stop()
+})
+
+describe('postcode page', () => {
+  test('should return a view with error message when redirected to with an error', async () => {
+    const options = {
+      method: 'GET',
+      url: '/postcode?error=true',
+      headers: {
+        cookie
+      }
+    }
+    const response = await server.inject(options)
+    expect(response.statusCode).toEqual(200)
+    expect(response.result).toMatch('This postcode does not appear to exist')
+  })
+
+  test('should redirect to search page when postcode submitted', async () => {
+    const { postOptions } = mockSearchOptions('NP18 3EZ', cookie)
+    const postResponse = await server.inject(postOptions)
+    expect(postResponse.statusCode).toEqual(STATUS_CODES.HTTP_STATUS_FOUND)
+    expect(postResponse.headers.location).toMatch(`/search?postcode=${encodeURIComponent('NP18 3EZ')}`)
+  })
+})
